+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - `ViewModel.$loadCleanData()` now returns the VM instance, to support call chaining.
 - `ListViewModel` now has shorthand properties for `$search`, `$filter`, `$orderBy`, and `$orderByDescending`.
 - `ListViewModel.$orderByToggle()` method added to cycle through ascending, descending, and no ordering for a field.
+- `ApiCaller.useRefResponse()` and `ApiClient.$useRefResponse()` methods added to enable System.Text.Json reference preservation handling, which can significantly reduce response payload sizes by deduplicating identical objects.
 - `c-admin-audit-log-page`: Added `userProp` prop to allow specifying a custom property name for user identification.
 - `c-admin-display`: Binary values now render as links that will download the value as a file, instead of only showing the length in bytes.
 - `c-datetime-picker`: Added prop `showTodayButton`